+++ tests/TestCase/Datasource/Paging/SimplePaginatorTest.php
@@ -11,27 +11,25 @@ declare(strict_types=1);
  *
  * @copyright     Copyright (c) Cake Software Foundation, Inc. (http://cakefoundation.org)
  * @link          http://cakephp.org CakePHP(tm) Project
- * @since         3.5.0
+ * @since         3.9.0
  * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\Test\TestCase\Datasource;
+namespace Cake\Test\TestCase\Datasource\Paging;
 
+use Cake\Core\Configure;
+use Cake\Datasource\Paging\SimplePaginator;
 use Cake\ORM\Entity;
-use Cake\TestSuite\TestCase;
 
-class PaginatorTest extends TestCase
+class SimplePaginatorTest extends PaginatorTest
 {
-    use PaginatorTestTrait;
+    public function setUp(): void
+    {
+        parent::setUp();
 
-    /**
-     * fixtures property
-     *
-     * @var array<string>
-     */
-    protected $fixtures = [
-        'core.Posts', 'core.Articles', 'core.Tags', 'core.ArticlesTags',
-        'core.Authors', 'core.AuthorsTags',
-    ];
+        Configure::write('App.namespace', 'TestApp');
+
+        $this->Paginator = new SimplePaginator();
+    }
 
     /**
      * test paginate() and custom find, to make sure the correct count is returned.
@@ -58,7 +56,7 @@ class PaginatorTest extends TestCase
 
         $pagingParams = $this->Paginator->getPagingParams();
         $this->assertSame(4, $pagingParams['PaginatorPosts']['current']);
-        $this->assertSame(4, $pagingParams['PaginatorPosts']['count']);
+        $this->assertNull($pagingParams['PaginatorPosts']['count']);
 
         $settings = ['finder' => 'published'];
         $result = $this->Paginator->paginate($table, [], $settings);
@@ -67,7 +65,7 @@ class PaginatorTest extends TestCase
 
         $pagingParams = $this->Paginator->getPagingParams();
         $this->assertSame(3, $pagingParams['PaginatorPosts']['current']);
-        $this->assertSame(3, $pagingParams['PaginatorPosts']['count']);
+        $this->assertNull($pagingParams['PaginatorPosts']['count']);
 
         $settings = ['finder' => 'published', 'limit' => 2, 'page' => 2];
         $result = $this->Paginator->paginate($table, [], $settings);
@@ -76,8 +74,8 @@ class PaginatorTest extends TestCase
 
         $pagingParams = $this->Paginator->getPagingParams();
         $this->assertSame(1, $pagingParams['PaginatorPosts']['current']);
-        $this->assertSame(3, $pagingParams['PaginatorPosts']['count']);
-        $this->assertSame(2, $pagingParams['PaginatorPosts']['pageCount']);
+        $this->assertNull($pagingParams['PaginatorPosts']['count']);
+        $this->assertSame(0, $pagingParams['PaginatorPosts']['pageCount']);
 
         $settings = ['finder' => 'published', 'limit' => 2];
         $result = $this->Paginator->paginate($table, [], $settings);
@@ -86,8 +84,8 @@ class PaginatorTest extends TestCase
 
         $pagingParams = $this->Paginator->getPagingParams();
         $this->assertSame(2, $pagingParams['PaginatorPosts']['current']);
-        $this->assertSame(3, $pagingParams['PaginatorPosts']['count']);
-        $this->assertSame(2, $pagingParams['PaginatorPosts']['pageCount']);
+        $this->assertNull($pagingParams['PaginatorPosts']['count']);
+        $this->assertSame(0, $pagingParams['PaginatorPosts']['pageCount']);
         $this->assertTrue($pagingParams['PaginatorPosts']['nextPage']);
         $this->assertFalse($pagingParams['PaginatorPosts']['prevPage']);
         $this->assertSame(2, $pagingParams['PaginatorPosts']['perPage']);
@@ -119,8 +117,8 @@ class PaginatorTest extends TestCase
 
         $result = $this->Paginator->getPagingParams()['PaginatorPosts'];
         $this->assertSame(2, $result['current']);
-        $this->assertSame(3, $result['count']);
-        $this->assertSame(2, $result['pageCount']);
+        $this->assertNull($result['count']);
+        $this->assertSame(0, $result['pageCount']);
         $this->assertTrue($result['nextPage']);
         $this->assertFalse($result['prevPage']);
     }
@@ -147,26 +145,7 @@ class PaginatorTest extends TestCase
 
         $this->assertSame(1, $pagingParams['PaginatorPosts']['start']);
         $this->assertSame(2, $pagingParams['PaginatorPosts']['end']);
-        $this->assertFalse($pagingParams['PaginatorPosts']['nextPage']);
-    }
-
-    /**
-     * test direction setting.
-     */
-    public function testPaginateDefaultDirection(): void
-    {
-        $settings = [
-            'PaginatorPosts' => [
-                'order' => ['Other.title' => 'ASC'],
-            ],
-        ];
-
-        $table = $this->getTableLocator()->get('PaginatorPosts');
-
-        $this->Paginator->paginate($table, [], $settings);
-        $pagingParams = $this->Paginator->getPagingParams();
-
-        $this->assertSame('Other.title', $pagingParams['PaginatorPosts']['sort']);
-        $this->assertNull($pagingParams['PaginatorPosts']['direction']);
+        // nextPage will be always true for SimplePaginator
+        $this->assertTrue($pagingParams['PaginatorPosts']['nextPage']);
     }
 }
