+++ tests/test_app/TestApp/Datasource/Paging/CustomPaginator.php
@@ -1,9 +1,9 @@
 <?php
 declare(strict_types=1);
 
-namespace TestApp\Datasource;
+namespace TestApp\Datasource\Paging;
 
-use Cake\Datasource\Paginator;
+use Cake\Datasource\Paging\Paginator;
 
 class CustomPaginator extends Paginator
 {
